requirements.txt
langchain
notebook
ipynb
huggingface_hub
langchain
openai
streamlit
python-dotenv
